fixed division

+++ math.c
@@ -12,6 +12,6 @@ int multiply(int a, int b){
     return a*b;
 }
 
-int divide(int a, int b){
+double divide(double a, double b){
     return a/b;
 }
