+++ math_library.h
@@ -4,6 +4,6 @@
 int sum(int a, int b);
 int sub(int a, int b);
 int multiply(int a, int b);
-int divide(int a, int b);
+double divide(double a, double b);
 
 #endif
